Factor EDTF validation out into a behavior
