+++ app/models/audio.rb
@@ -4,6 +4,7 @@
 class Audio < ActiveFedora::Base
   include ::Hyrax::WorkBehavior
   include ::OregonDigital::WorkBehavior
+  include ::OregonDigital::ValidatesEDTFBehavior
 
   self.indexer = AudioIndexer
   # Change this to restrict which works can be added as a child.
