+++ app/models/document.rb
@@ -4,6 +4,7 @@
 class Document < ActiveFedora::Base
   include ::Hyrax::WorkBehavior
   include ::OregonDigital::WorkBehavior
+  include ::OregonDigital::ValidatesEDTFBehavior
 
   self.indexer = DocumentIndexer
   # Change this to restrict which works can be added as a child.
