+++ app/models/image.rb
@@ -4,6 +4,7 @@
 class Image < ActiveFedora::Base
   include ::Hyrax::WorkBehavior
   include ::OregonDigital::WorkBehavior
+  include ::OregonDigital::ValidatesEDTFBehavior
 
   self.indexer = ImageIndexer
   # Change this to restrict which works can be added as a child.
